"Refactor code to improve organization and naming conventions
This commit performs several improvements to the organization and naming conventions of the code.

By moving certain parts of code to more suitable locations such as moving command lines to `DayLight.Subclasses.csproj` and renaming `DataType` to `MessageType` to bring more clarity on its functionality, we have improved the overall maintainability and readability of the code.

Also, certain patches have been applied to increase the privacy and security of certain parts of the system described in the `DayLight.Core` module. Now, all the methods in `Network.cs` have been made `private` to avoid unauthorized modifications. This will raise the integrity of the data and its associated operations.

In `DayLight.Core API`, the renaming  has been done to suit the semantics of the file and its functionalities.

No functionality changes were performed by these commits, it was purely focused on improving the codebase usability and organization."