Add new Neuron projects and improve plugin logging
Added new "Neuron" and "Neuron.Bootstrap" projects to the DayLight solution. Relevance of these new additions is to expand functionality and improve modularity of the codebase. Refactored all plugin files to have a consistent naming pattern and to improve clarity. This appears in more informative plugin load logging. Updated how the "CustomRaCategory" is managed and substituting "CategoryManager" with it to improve code legibility. Moreover, code was updated to use the newer dictionary initialization syntax for readability.