Refactor database player fetching for efficiency
Refactored the way the player is fetched from the database across several files to improve efficiency and readability. The method to fetch a player from the database was shifted to AdvancedPlayer class and it now automatically updates the player on any changes in player properties. It uses a built-in PropertyChanged event in .NET to listen for changes.

This reduces code duplication and ensures that the player instance in the database is always up-to-date.

Also, a method for adding and removing players was fine-tuned to handle situations where player-do-not-track flag is set. If a player has this flag set, they are not added to the database instead of being added and then removed immediately.

Bugs that came from the misspelled word "Achivement" were fixed by replacing it to the correct spelling "Achievement".